feat: use stdlib path glob
On-behalf-of: SAP <matthias.bruns@sap.com>
